// frontend/src/types/contact.ts
export interface Contact {
  _id?: string;

  companyName: string;

  address: string;

  phone1: string;

  phone2: string;

  email: string;

  whatsapp: string;

  googleMap: string;

  facebook: string;

  instagram: string;

  youtube: string;

  linkedin: string;

  officeHours: string;

  copyright: string;

  isActive: boolean;

  createdAt?: string;

  updatedAt?: string;
}